@@ -10,50 +10,56 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^15.0.0",
-    "@angular/common": "^15.0.0",
-    "@angular/compiler": "^15.0.0",
-    "@angular/core": "^15.0.0",
-    "@angular/forms": "^15.0.0",
-    "@angular/platform-browser": "^15.0.0",
-    "@angular/platform-browser-dynamic": "^15.0.0",
-    "@angular/router": "^15.0.0",
-    "@fontsource/poppins": "^4.5.10",
-    "@ngx-translate/core": "^14.0.0",
-    "@ngx-translate/http-loader": "^7.0.0",
+    "@angular/animations": "18.2.13",
+    "@angular/common": "18.2.13",
+    "@angular/compiler": "18.2.13",
+    "@angular/core": "18.2.13",
+    "@angular/forms": "18.2.13",
+    "@angular/platform-browser": "18.2.13",
+    "@angular/platform-browser-dynamic": "18.2.13",
+    "@angular/router": "18.2.13",
+    "@fontsource/poppins": "^5.0.14",
+    "@ngx-cfng-core-layout/header": "file:dist/ngx-cfng-core-layout/header",
+    "@ngx-cfng-core-layout/sidebar": "file:dist/ngx-cfng-core-layout/sidebar",
+    "@ngx-cfng-core-modal/dialog": "file:dist/ngx-cfng-core-modal/dialog",
+    "@ngx-translate/core": "^15.0.0",
+    "@ngx-translate/http-loader": "^8.0.0",
     "@types/jszip": "^3.4.0",
-    "crypto-js": "^4.1.1",
-    "date-fns": "^2.29.3",
-    "inputmask": "^5.0.8",
+    "crypto-js": "^4.2.0",
+    "date-fns": "^3.6.0",
+    "inputmask": "^5.0.9",
     "jsbarcode": "^3.11.5",
     "jszip": "^3.10.1",
     "jwt-decode": "^3.1.2",
-    "leaflet": "^1.9.3",
-    "ng2-pdf-viewer": "^9.1.5",
-    "ngx-extended-pdf-viewer": "^16.2.16",
-    "ngx-mpfn-dev-cmp-lib": "^0.4.33",
+    "leaflet": "^1.9.4",
+    "ngx-cfng-core-firma-digital": "file:dist/ngx-cfng-core-firma-digital",
+    "ngx-cfng-core-lib": "file:dist/ngx-cfng-core-lib",
+    "ngx-extended-pdf-viewer": "^21.3.5",
+    "ngx-joyride": "^2.5.0",
+    "ngx-mpfn-dev-cmp-lib": "^0.4.36",
     "ngx-mpfn-dev-icojs-mpfn": "^1.0.1",
-    "ngx-mpfn-dev-icojs-regular": "^1.0.13",
+    "ngx-mpfn-dev-icojs-regular": "^1.0.18",
     "pdf-lib": "^1.17.1",
-    "primeflex": "^3.3.0",
+    "pdfjs-dist": "^5.3.31",
+    "primeflex": "^3.3.1",
     "primeicons": "^7.0.0",
-    "primeng": "^15.0.1",
-    "rxjs": "~7.5.0",
+    "primeng": "^17.18.12",
+    "rxjs": "~7.8.0",
     "tslib": "^2.3.0",
-    "zone.js": "~0.12.0"
+    "zone.js": "~0.14.3"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^15.0.4",
-    "@angular/cli": "~15.0.4",
-    "@angular/compiler-cli": "^15.0.0",
-    "@types/jasmine": "~4.3.0",
-    "@types/leaflet": "^1.9.3",
-    "jasmine-core": "~4.5.0",
+    "@angular-devkit/build-angular": "18.2.3",
+    "@angular/cli": "18.2.3",
+    "@angular/compiler-cli": "18.2.13",
+    "@types/jasmine": "~5.1.0",
+    "@types/leaflet": "^1.9.12",
+    "jasmine-core": "~5.1.0",
     "karma": "~6.4.0",
-    "karma-chrome-launcher": "~3.1.0",
+    "karma-chrome-launcher": "~3.2.0",
     "karma-coverage": "~2.2.0",
     "karma-jasmine": "~5.1.0",
-    "karma-jasmine-html-reporter": "~2.0.0",
-    "typescript": "~4.8.2"
+    "karma-jasmine-html-reporter": "~2.1.0",
+    "typescript": "5.5.2"
   }
 }
