@@ -6,14 +6,17 @@ import { routes } from './app/app-routes';
 import { PrimeNGConfig } from 'primeng/api';
 import { TranslateLoader, TranslateModule, TranslateService } from '@ngx-translate/core';
 import { HashLocationStrategy, LocationStrategy } from '@angular/common';
-import { HTTP_INTERCEPTORS, HttpClient, HttpClientModule } from '@angular/common/http';
+import { HTTP_INTERCEPTORS, HttpClient, provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
 import { RequestInterceptor } from '@core/interceptors/request.interceptor';
 import { TranslateHttpLoader } from '@ngx-translate/http-loader';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 
+
 export function HttpLoaderFactory(httpClient: HttpClient) {
   return new TranslateHttpLoader(httpClient);
 }
+
+
 
 bootstrapApplication(AppComponent, {
   providers: [
@@ -21,8 +24,8 @@ bootstrapApplication(AppComponent, {
     TranslateService,
     { provide: LocationStrategy, useClass: HashLocationStrategy },
     { provide: HTTP_INTERCEPTORS, useClass: RequestInterceptor, multi: true },
+    provideHttpClient(withInterceptorsFromDi()),
     importProvidersFrom(RouterModule.forRoot(routes)),
-    importProvidersFrom(HttpClientModule),
     importProvidersFrom(BrowserAnimationsModule),
     importProvidersFrom(
       TranslateModule.forRoot({
@@ -33,7 +36,8 @@ bootstrapApplication(AppComponent, {
         },
       })
     ),
+
   ],
 }).catch((err) =>
   console.error(err)
-); 
+);
