@@ -7,19 +7,16 @@ import { DenunciaRegistradaComponent } from '@modules/complaint/components/denun
 import { ConsultCaseComponent } from '@modules/tracing/components/consult-case/consult-case.component';
 import { CompletedProcessComponent } from '@modules/tracing/components/completed-process/completed-process.component';
 import { AppendDocumentComponent } from '@modules/append/components/append-document/append-document.component';
-import {
-  DocumentPreviewComponent
-} from "@modules/append/components/completed-process/document-preview/document-preview.component";
-import {
-  DocumentRegisteredComponent
-} from "@modules/append/components/completed-process/document-registered/document-registered.component";
+import { DocumentPreviewComponent } from "@modules/append/components/completed-process/document-preview/document-preview.component";
+import { DocumentRegisteredComponent } from "@modules/append/components/completed-process/document-registered/document-registered.component";
 import { SpecialtyComponent } from '@modules/complaint/components/specialty/specialty.component';
-import {VerificationIdPeruComponent} from "@shared/components/verificationIdPeru/verification-id-peru.component";
+import { VerificationIdPeruComponent } from "@shared/components/verificationIdPeru/verification-id-peru.component";
 
 export const routes: Routes = [
   {
     path: '',
-    loadComponent: () => import('./modules/main/main.component').then(m => m.MainComponent)
+    loadComponent: () => import('./modules/main/main.component').then(m => m.MainComponent),
+    pathMatch: 'full'
   },
   {
     path: 'realizar-denuncia',
@@ -27,7 +24,9 @@ export const routes: Routes = [
     children: [
       { path: '', redirectTo: 'verificacion', pathMatch: 'full' },
       { path: 'verificacion', component: VerificationComponent },
+
       { path: 'datos-personales', component: PersonalDataComponent },
+
       { path: 'datos-especialidad', component: SpecialtyComponent },
       { path: 'datos-denuncia', component: ComplaintDataComponent },
       { path: 'confirmacion', component: ConfirmationComponent, },
@@ -63,4 +62,9 @@ export const routes: Routes = [
     path: 'verificacion-id-peru',
     component: VerificationIdPeruComponent
   },
+  {
+    path: '**',
+    redirectTo: '',
+    pathMatch: 'full'
+  }
 ]
