@@ -1,2 +1,2 @@
-export type DomainType = 'auth' | 'maestros' | 'persona' | 'mesa' | 'documento' | 'sunat' | 'repositorio' | 'denuncia' | 'electronica';
+export type DomainType = 'auth' | 'maestros' | 'persona' | 'mesa' | 'documento' | 'sunat' | 'repositorio' | 'denuncia' | 'electronica' | 'ai';
 
